change schema.

+++ database/sql/main.sql
@@ -30,12 +30,12 @@ CREATE TABLE `user_business` (
 DROP TABLE IF EXISTS `business`;
 
 CREATE TABLE `business` (
-    `id` BIGINT unsigned NOT NULL,
+    `id` INTEGER unsigned NOT NULL,
     `prefecture` INTEGER unsigned NOT NULL,
     `name` VARCHAR(191) NOT NULL,
-    `price_base` INTEGER unsigned NOT NULL,
-    `price_level2` INTEGER unsigned NOT NULL,
-    `price_level3` INTEGER unsigned NOT NULL,
+    `price_base` BIGINT unsigned NOT NULL,
+    `price_level2` BIGINT unsigned NOT NULL,
+    `price_level3` BIGINT unsigned NOT NULL,
     `return_rate_base` INTEGER unsigned NOT NULL,
     `return_rate_level2` INTEGER unsigned NOT NULL,
     `return_rate_level3` INTEGER unsigned NOT NULL,
